+++ cmd/launcher/svc_windows.go
@@ -17,8 +17,8 @@ import (
 	"github.com/kolide/launcher/pkg/autoupdate"
 	"github.com/kolide/launcher/pkg/contexts/ctxlog"
 	"github.com/kolide/launcher/pkg/launcher"
-	"github.com/kolide/launcher/pkg/log/debuglogger"
 	"github.com/kolide/launcher/pkg/log/eventlog"
+	"github.com/kolide/launcher/pkg/log/locallogger"
 	"github.com/kolide/launcher/pkg/log/teelogger"
 	"github.com/pkg/errors"
 	"golang.org/x/sys/windows/svc"
@@ -51,9 +51,9 @@ func runWindowsSvc(args []string) error {
 		os.Exit(1)
 	}
 
-	// Create a local, debug, logger. This logs to a known path
+	// Create a local logger. This logs to a known path, and aims to help diagnostics
 	if opts.RootDirectory != "" {
-		logger = teelogger.New(logger, debuglogger.NewKitLogger(filepath.Join(opts.RootDirectory, "debug.log")))
+		logger = teelogger.New(logger, locallogger.NewKitLogger(filepath.Join(opts.RootDirectory, "debug.log")))
 	}
 
 	// Now that we've parsed the options, let's set a filter on our logger
